DDL for tables in Database
I went ahead and created a sql file that should create the tables we
need based on the ERD Mike showed everyone in class. This is incomplete
because I don’t have my SQL workbench and can only work off the
screenshots. Feel free to change or modify it if you see something that
needs to be changed or added.